+++ mindsdb/integrations/handlers/gcs_handler/connection_args.py
@@ -12,11 +12,13 @@ connection_args = OrderedDict(
     service_account_keys={
         'type': ARG_TYPE.PATH,
         'description': 'Path to the service account JSON file',
+        'label': 'Path to the service account JSON file',
         'secret': True
     },
     service_account_json={
         'type': ARG_TYPE.DICT,
         'description': 'Content of service account JSON file',
+        'label': 'Content of service account JSON file',
         'secret': True
     }
 )
