+++ mindsdb/integrations/handlers/gcs_handler/gcs_handler.py
@@ -19,45 +19,9 @@ from mindsdb.integrations.libs.response import (
     RESPONSE_TYPE
 )
 
-from mindsdb.integrations.libs.api_handler import APIResource, APIHandler
-from mindsdb.integrations.utilities.sql_utils import FilterCondition
+from mindsdb.integrations.libs.api_handler import APIHandler
 
 logger = log.getLogger(__name__)
-
-
-class ListFilesTable(APIResource):
-
-    def list(self,
-             targets: List[str] = None,
-             conditions: List[FilterCondition] = None,
-             *args, **kwargs) -> pd.DataFrame:
-
-        tables = self.handler._get_tables()
-        data = []
-        for path in tables:
-            path = path.replace('`', '')
-            item = {
-                'path': path,
-                'name': path[path.rfind('/') + 1:],
-                'extension': path[path.rfind('.') + 1:]
-            }
-
-            data.append(item)
-
-        return pd.DataFrame(data=data, columns=self.get_columns())
-
-    def get_columns(self) -> List[str]:
-        return ["path", "name", "extension", "content"]
-
-
-class FileTable(APIResource):
-
-    def list(self, targets: List[str] = None, table_name=None, *args, **kwargs) -> pd.DataFrame:
-        return self.handler._read_as_table(table_name)
-
-    def add(self, data, table_name=None):
-        df = pd.DataFrame(data)
-        return self.handler._add_data_to_table(table_name, df)
 
 
 class GcsHandler(APIHandler):
@@ -82,7 +46,6 @@ class GcsHandler(APIHandler):
         self.connection_data = connection_data
         self.kwargs = kwargs
         self.is_select_query = False
-        self.table_name = None
         self.service_account_json = None
         self.connection = None
